+++ apps/readme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # readme-widget-hub
 
+## 4.1.0
+
+### Minor Changes
+
+- 小部件支持用法描述
+
+### Patch Changes
+
+- Updated dependencies
+  - @readme-widget-hub/render@2.2.0
+  - @readme-widget-hub/widget@2.3.0
+  - @readme-widget-hub/meta@2.1.0
+  - @readme-widget-hub/manager@2.0.4
+
 ## 4.0.3
 
 ### Patch Changes