+++ packages/render/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @readme-widget-hub/render
 
+## 2.2.0
+
+### Minor Changes
+
+- 小部件支持用法描述
+
+### Patch Changes
+
+- Updated dependencies
+  - @readme-widget-hub/widget@2.3.0
+  - @readme-widget-hub/meta@2.1.0
+
 ## 2.1.0
 
 ### Minor Changes